requirements.txt
pywebview==4.4.1
google-cloud-vision==3.7.0
Pillow==10.2.0
openpyxl==3.1.2
python-dotenv==1.0.0
PyMuPDF==1.23.8
